feat: make file to creater server & import to server.ts
